test -1*-1 and maxModulus() (should make sure that nextprime(maxModulus()) fails \!\!\!

+++ trunk/linbox/tests/test-smith-form-binary.C
@@ -229,8 +229,7 @@ int main(int argc, char** argv)
 
 		RandomDenseStream<Ring> s1 (R, n, iterations);
 
-		// typedef Modular<int32_t> Field;
-		typedef Modular<double> Field;
+		typedef Modular<int32_t> Field;
 		typedef RationalSolver<Ring, Field, LinBox::RandomPrimeIterator> Solver;
 		typedef LastInvariantFactor<Ring, Solver> LIF;
 		typedef OneInvariantFactor<Ring, LIF, SCompose, RandomMatrix>  OIF;
